% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{amapping}
\alias{amapping}
\title{Apply a Mapping}
\usage{
amapping(x, ...)
}
\arguments{
\item{x}{The values to apply the \code{\link{mapping}} to.}

\item{...}{Passed to \code{\link{mapping}}.}
}
\value{
The values returned by \code{\link{mapping}}.
}
\description{
Apply a Mapping
}
