% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{as.data.frame.mapping}
\alias{as.data.frame.mapping}
\title{Convert a mapping to \code{data.frame}}
\usage{
\method{as.data.frame}{mapping}(x, ...)
}
\arguments{
\item{x}{A \code{\link{mapping}}}

\item{...}{Ignored}
}
\value{
A \code{data.frame}.
}
\description{
The resulting \code{data.frame} has 2 columns: \code{mapsfrom}, and \code{mapsto}.
}
