% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{cut_mapping}
\alias{cut_mapping}
\title{Mapping from continuous to categorical}
\usage{
cut_mapping(..., to = NULL, na = NA, ch.as.fact = TRUE)
}
\arguments{
\item{...}{Passed to \code{\link[base:cut]{cut()}}.}

\item{to}{Passed to \code{\link[=mapping]{mapping()}}.}

\item{na}{Passed to \code{\link[=mapping]{mapping()}}.}

\item{ch.as.fact}{Passed to \code{\link[=mapping]{mapping()}}.}
}
\value{
A function that cuts a \code{numeric} vector and maps the result.
}
\description{
Mapping from continuous to categorical
}
\examples{
x <- c(0, 10, 20, 30, Inf)
m <- cut_mapping(x, right=FALSE,
    to=c("0 to <10", "10 to <20", "20 to <30", ">= 30"))
print(m)
m(c(5, 27, 3, 10, 99))
}
