% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{print.mapping}
\alias{print.mapping}
\title{Print a mapping}
\usage{
\method{print}{mapping}(x, ...)
}
\arguments{
\item{x}{A \code{\link{mapping}}}

\item{...}{Ignored}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print a mapping
}
