% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{remap}
\alias{remap}
\title{Re-map a variable}
\usage{
remap(x, ...)
}
\arguments{
\item{x}{The values to apply the \code{\link{mapping}} to.}

\item{...}{Passed to \code{\link{mapping}}.}
}
\value{
The values returned by \code{\link{mapping}}.
}
\description{
Re-map a variable
}
