% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{domain}
\alias{domain}
\alias{codomain}
\title{Domain and codomain of a mapping.}
\usage{
domain(x)

codomain(x)
}
\arguments{
\item{x}{A \code{\link{mapping}}}
}
\description{
Domain and codomain of a mapping.
}
