% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{remap}
\alias{remap}
\title{Re-map a variable}
\usage{
remap(x, ...)
}
\arguments{
\item{x}{The values to apply the \code{\link{mapping}} to.}

\item{...}{Passed to \code{\link[=mapping]{mapping()}}.}
}
\value{
The values returned by calling the \code{\link{mapping}} function.
}
\description{
Apply a mapping to a vector directly. The mapping is temporary and not saved.
}
\examples{
x <- c("A", "B", "A")
remap(x, c(A=0, B=1))
}
